# 55% speedup for HashWithIndifferentAccess.new when no args provided